+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/internal/JavaLanguageProcessor.java
@@ -4,6 +4,10 @@
 
 package net.sourceforge.pmd.lang.java.internal;
 
+import java.io.File;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.List;
 import java.util.Objects;
 
@@ -11,6 +15,7 @@ import org.checkerframework.checker.nullness.qual.NonNull;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import net.sourceforge.pmd.lang.JvmLanguagePropertyBundle;
 import net.sourceforge.pmd.lang.LanguageVersionHandler;
 import net.sourceforge.pmd.lang.ast.Parser;
 import net.sourceforge.pmd.lang.impl.BatchLanguageProcessor;
@@ -45,10 +50,12 @@ public class JavaLanguageProcessor extends BatchLanguageProcessor<JavaLanguagePr
     private final JavaParser parserWithoutProcessing;
     private final boolean firstClassLombok;
     private TypeSystem typeSystem;
+    private AuxClasspathLoader auxClasspathLoader;
 
-    public JavaLanguageProcessor(JavaLanguageProperties properties, TypeSystem typeSystem) {
+    public JavaLanguageProcessor(JavaLanguageProperties properties) {
         super(properties);
-        this.typeSystem = typeSystem;
+
+        initTypeSystem(properties);
 
         String suppressMarker = properties.getSuppressMarker();
         this.parser = new JavaParser(suppressMarker, this, true);
@@ -56,9 +63,31 @@ public class JavaLanguageProcessor extends BatchLanguageProcessor<JavaLanguagePr
         this.firstClassLombok = properties.getProperty(JavaLanguageProperties.FIRST_CLASS_LOMBOK);
     }
 
-    public JavaLanguageProcessor(JavaLanguageProperties properties) {
-        this(properties, TypeSystem.usingClassLoaderClasspath(properties.getAnalysisClassLoader()));
-        LOG.debug("Using analysis classloader: {}", properties.getAnalysisClassLoader());
+    private void initTypeSystem(JavaLanguageProperties properties) {
+        ClassLoader externallyConfiguratedClassLoader = properties.getExternalClassLoader();
+        if (externallyConfiguratedClassLoader != null) {
+            LOG.debug("Using externally configured classloader as analysis classloader: {}", externallyConfiguratedClassLoader);
+            this.typeSystem = TypeSystem.usingClassLoaderClasspath(externallyConfiguratedClassLoader);
+        } else {
+            String auxClasspath = properties.getProperty(JvmLanguagePropertyBundle.AUX_CLASSPATH);
+
+            if (!auxClasspath.contains("jrt-fs.jar") && !auxClasspath.contains("rt.jar")) {
+                // TODO use AuxClasspathUtil.getPlatformClasspath() from #6841
+                Path jrtFsJar = Paths.get(System.getProperty("java.home"), "lib", "jrt-fs.jar"); // Java 11+
+                Path rtJar = Paths.get(System.getProperty("java.home"), "jre", "lib", "rt.jar"); // Java 8
+                if (Files.isRegularFile(jrtFsJar)) {
+                    LOG.debug("Adding current JVM runtime classes from {}", jrtFsJar);
+                    auxClasspath += File.pathSeparator + jrtFsJar;
+                } else if (Files.isRegularFile(rtJar)) {
+                    LOG.debug("Adding current JVM runtime classes from {}", rtJar);
+                    auxClasspath += File.pathSeparator + rtJar;
+                }
+            }
+            LOG.debug("Using auxClasspath as analysis classloader: {}", auxClasspath);
+
+            this.auxClasspathLoader = new AuxClasspathLoader(auxClasspath);
+            this.typeSystem = TypeSystem.usingClasspath(auxClasspathLoader);
+        }
     }
 
     @Override
@@ -139,6 +168,9 @@ public class JavaLanguageProcessor extends BatchLanguageProcessor<JavaLanguagePr
     @Override
     public void close() throws Exception {
         this.typeSystem.logStats();
+        if (this.auxClasspathLoader != null) {
+            this.auxClasspathLoader.close();
+        }
         super.close();
     }
 }