+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/symbols/ClassLoadingChildFirstTest.java
@@ -9,14 +9,18 @@ import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.hasProperty;
 import static org.hamcrest.Matchers.hasSize;
 
+import java.io.File;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
-import net.sourceforge.pmd.PMDConfiguration;
 import net.sourceforge.pmd.internal.util.ClasspathClassLoader;
+import net.sourceforge.pmd.lang.JvmLanguagePropertyBundle;
+import net.sourceforge.pmd.lang.java.JavaLanguageModule;
+import net.sourceforge.pmd.lang.java.internal.AuxClasspathLoader;
+import net.sourceforge.pmd.lang.java.internal.JavaLanguageProperties;
 import net.sourceforge.pmd.lang.java.types.JClassType;
 import net.sourceforge.pmd.lang.java.types.TypeSystem;
 
@@ -38,20 +42,25 @@ class ClassLoadingChildFirstTest {
      * </p>
      */
     @Test
-    void testClassLoading() {
+    void testClassLoading() throws Exception {
         Path file = Paths.get("src/test/resources",
                 getClass().getPackage().getName().replace('.', '/'),
                 "custom_java_lang.jar");
+        Path jrtFsJar = Paths.get(System.getProperty("java.home"), "lib", "jrt-fs.jar");
+        String classpath = file.toAbsolutePath() + File.pathSeparator + jrtFsJar;
 
-        PMDConfiguration config = new PMDConfiguration();
-        config.prependAuxClasspath(file.toAbsolutePath().toString());
+        JavaLanguageProperties languageProperties = (JavaLanguageProperties) JavaLanguageModule.getInstance().newPropertyBundle();
+        languageProperties.setProperty(JvmLanguagePropertyBundle.AUX_CLASSPATH, classpath);
 
-        TypeSystem typeSystem = TypeSystem.usingClassLoaderClasspath(config.getClassLoader());
+        try (AuxClasspathLoader auxClasspathLoader = new AuxClasspathLoader(classpath)) {
+            TypeSystem typeSystem = TypeSystem.usingClasspath(auxClasspathLoader);
 
-        JClassType voidClass = typeSystem.BOXED_VOID;
-        List<JMethodSymbol> declaredMethods = voidClass.getSymbol().getDeclaredMethods();
-        assertThat(declaredMethods, hasSize(1));
-        assertThat(declaredMethods.get(0), hasProperty("simpleName", equalTo("customMethodOnJavaLangVoid")));
+            JClassType voidClass = typeSystem.BOXED_VOID;
+            List<JMethodSymbol> declaredMethods = voidClass.getSymbol().getDeclaredMethods();
+            assertThat(declaredMethods, hasSize(1));
+            assertThat(declaredMethods.get(0), hasProperty("simpleName", equalTo("customMethodOnJavaLangVoid")));
+
+        }
     }
 
 }